Moved multipole work to the tree, began optimising neighbour loading
