+++ src/Headers/BruteForceTree.h
@@ -98,7 +98,7 @@ class BruteForceTree : public Tree<ndim,ParticleType,TreeCell>
 
   // Constructor and destructor
   //-----------------------------------------------------------------------------------------------
-  BruteForceTree(int, FLOAT, FLOAT, FLOAT, string, string, const DomainBox<ndim>&,
+  BruteForceTree(int, FLOAT, FLOAT, FLOAT, string, multipole_method, const DomainBox<ndim>&,
 		  	  	 const ParticleTypeRegister& reg, const bool);
   ~BruteForceTree() ;
