+++ src/Headers/SphNeighbourSearch.h
@@ -113,7 +113,6 @@ protected:
   using HydroTree<ndim,ParticleType>::activepartbuf;
   using HydroTree<ndim,ParticleType>::allocated_buffer;
   using HydroTree<ndim,ParticleType>::box;
-  using HydroTree<ndim,ParticleType>::cellbuf;
   using HydroTree<ndim,ParticleType>::gravity_mac;
   using HydroTree<ndim,ParticleType>::kernp;
   using HydroTree<ndim,ParticleType>::kernrange;
@@ -144,7 +143,7 @@ protected:
   //-----------------------------------------------------------------------------------------------
   SphTree(string tree_t, int _Nleafmax, int _Nmpi, int _pruning_level_min, int _pruning_level_max,
           FLOAT _thetamaxsqd, FLOAT _kernrange, FLOAT _macerror,
-          string _gravity_mac, string _multipole,
+          string _gravity_mac, multipole_method _multipole,
           DomainBox<ndim> *_box, SmoothingKernel<ndim> *_kern, CodeTiming *_timing,
           ParticleTypeRegister& types) :
     HydroTree<ndim,ParticleType>(tree_t, _Nleafmax, _Nmpi, _pruning_level_min, _pruning_level_max,
@@ -183,7 +182,6 @@ private:
   using SphTree<ndim,ParticleType>::activepartbuf;
   using SphTree<ndim,ParticleType>::allocated_buffer;
   using SphTree<ndim,ParticleType>::box;
-  using SphTree<ndim,ParticleType>::cellbuf;
   using SphTree<ndim,ParticleType>::gravity_mac;
   using SphTree<ndim,ParticleType>::kernp;
   using SphTree<ndim,ParticleType>::kernrange;
@@ -212,7 +210,7 @@ private:
 
 
   //-----------------------------------------------------------------------------------------------
-  GradhSphTree(string, int, int, int, int, FLOAT, FLOAT, FLOAT, string, string,
+  GradhSphTree(string, int, int, int, int, FLOAT, FLOAT, FLOAT, string, multipole_method,
                DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *,
                ParticleTypeRegister& types);
   virtual ~GradhSphTree();
@@ -242,7 +240,6 @@ class SM2012SphTree: public SphTree<ndim,ParticleType>
   using SphTree<ndim,ParticleType>::activepartbuf;
   using SphTree<ndim,ParticleType>::allocated_buffer;
   using SphTree<ndim,ParticleType>::box;
-  using SphTree<ndim,ParticleType>::cellbuf;
   using SphTree<ndim,ParticleType>::gravity_mac;
   using SphTree<ndim,ParticleType>::kernp;
   using SphTree<ndim,ParticleType>::kernrange;
@@ -270,7 +267,7 @@ class SM2012SphTree: public SphTree<ndim,ParticleType>
 
 
   //-----------------------------------------------------------------------------------------------
-  SM2012SphTree(string, int, int, int, int, FLOAT, FLOAT, FLOAT, string, string,
+  SM2012SphTree(string, int, int, int, int, FLOAT, FLOAT, FLOAT, string, multipole_method,
                 DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *,
                 ParticleTypeRegister&);
