+++ src/MeshlessFV/MeshlessFVSimulation.cpp
@@ -218,10 +218,24 @@ void MeshlessFVSimulation<ndim>::ProcessParameters(void)
   //-----------------------------------------------------------------------------------------------
   string tree_type = stringparams["neib_search"] ;
 
+  multipole_method multipole ;
+  if (stringparams["multipole"] == "monopole")
+    multipole = monopole ;
+  else if (stringparams["multipole"] == "quadrupole")
+    multipole = quadrupole ;
+  else if (stringparams["multipole"] == "fast_monopole")
+    multipole = fast_monopole ;
+  else if (stringparams["multipole"] == "fast_quadrupole")
+    multipole = fast_quadrupole ;
+  else {
+    string message = "Multipole type not recognised.";
+    ExceptionHandler::getIstance().raise(message);
+  }
+
   mfvneib = new MeshlessFVTree<ndim,MeshlessFVParticle>
   (tree_type, intparams["Nleafmax"], Nmpi, intparams["pruning_level_min"], intparams["pruning_level_max"],
       floatparams["thetamaxsqd"], hydro->kernp->kernrange, floatparams["macerror"],
-      stringparams["gravity_mac"], stringparams["multipole"], &simbox, mfv->kernp, timing, mfv->types);
+      stringparams["gravity_mac"], multipole, &simbox, mfv->kernp, timing, mfv->types);
 
   neib = mfvneib ;
   // Here I do a horrible hack to get at the underlying tree, needed for the dust.
