+++ src/Tree/KDTree.cpp
@@ -51,7 +51,7 @@ using namespace std;
 template <int ndim, template<int> class ParticleType, template<int> class TreeCell>
 KDTree<ndim,ParticleType,TreeCell>::KDTree(int Nleafmaxaux, FLOAT thetamaxsqdaux,
                                            FLOAT kernrangeaux, FLOAT macerroraux,
-                                           string gravity_mac_aux, string multipole_aux,
+                                           string gravity_mac_aux, multipole_method multipole_aux,
                                            const DomainBox<ndim>& domain,
                                   		   const ParticleTypeRegister& reg,
 										   const bool IAmPruned):
@@ -727,7 +727,7 @@ void KDTree<ndim,ParticleType,TreeCell>::StockCellProperties
 
 
   const bool need_quadrupole_moments =
-      multipole == "quadrupole" || multipole == "fast_quadrupole" || gravity_mac == eigenmac ;
+      multipole == quadrupole || multipole == fast_quadrupole || gravity_mac == eigenmac ;
 
   // Zero all summation variables for all cells
   if ((cell.level==ltot&&stock_leaf) || cell.copen != -1 ) {
